+++ Core/System/Linq/Expressions/Interpreter/SubInstruction.net30.cs
@@ -81,7 +81,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int64)unchecked((Int64)l - (Int64)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((Int64)l - (Int64)r);
                 }
                 frame.StackIndex--;
                 return +1;
@@ -119,7 +119,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt32)unchecked((UInt32)l - (UInt32)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((UInt32)l - (UInt32)r);
                 }
                 frame.StackIndex--;
                 return +1;
@@ -138,7 +138,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt64)unchecked((UInt64)l - (UInt64)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((UInt64)l - (UInt64)r);
                 }
                 frame.StackIndex--;
                 return +1;
@@ -157,7 +157,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Single)((Single)l - (Single)r);
+                    frame.Data[frame.StackIndex - 2] = (Single)l - (Single)r;
                 }
                 frame.StackIndex--;
                 return +1;
@@ -296,7 +296,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int64)checked((Int64)l - (Int64)r);
+                    frame.Data[frame.StackIndex - 2] = checked((Int64)l - (Int64)r);
                 }
                 frame.StackIndex--;
                 return +1;
@@ -334,7 +334,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt32)checked((UInt32)l - (UInt32)r);
+                    frame.Data[frame.StackIndex - 2] = checked((UInt32)l - (UInt32)r);
                 }
                 frame.StackIndex--;
                 return +1;
@@ -372,7 +372,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Single)((Single)l - (Single)r);
+                    frame.Data[frame.StackIndex - 2] = (Single)l - (Single)r;
                 }
                 frame.StackIndex--;
                 return +1;
